Added spreadsheet with Ghisalberti's Threshold
The threshold Reynolds number with data from Ghisalberti and Nepf (2002) (heneforth GN2002) is calculated to be compared with our results. The data used from table 1 of GN2002 reports the flow rate, the grass height, and water column depth. This data can be converted in terms of our parameters into a scale for velocity U_0. Assuming a turbulent viscosity of 1e-4 m^2/s, gives a Reynolds number for each case. Reynolds numbers below 200 are reported to be stable, while the ones above 200 are reported to be unstable by GN2002. We can plot these Reynolds numbers in our threshold plot now.