Improved discussion and conclusion
The comparison with frequency is dealt with first so as to combine the discussion of the threshold into one coherent section. The note about disagreement due to variation of grass density and eddy viscosity with depth is moved to a paragraph on limitations near the end of the paper. And the conclusion is improved to summarize the paper.